Refactor layout components to use a shared layout structure for better code reuse and organization
