+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/profiles/layout.tsx
@@ -1,28 +1,3 @@
-import { DashboardContentBody } from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-body";
-import { DashboardContentSection } from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-top-section";
+import SharedLayout from "~/app/(dashboard)/dashboard/_components/layout-shared";
 
-interface LayoutProps {
-  children: React.ReactNode;
-  left: React.ReactNode;
-  center: React.ReactNode;
-  right: React.ReactNode;
-}
-
-export default function Content({
-  children,
-  left,
-  center,
-  right,
-}: LayoutProps) {
-  return (
-    <>
-      <div className="grid auto-rows-min gap-4 md:grid-cols-3">
-        <DashboardContentSection>{left}</DashboardContentSection>
-        <DashboardContentSection>{center}</DashboardContentSection>
-        <DashboardContentSection>{right}</DashboardContentSection>
-      </div>
-
-      <DashboardContentBody>{children}</DashboardContentBody>
-    </>
-  );
-}
+export default SharedLayout;
