+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/layout_alt.tsx
